requirements.txt
matplotlib
Pillow
scikit-learn
numpy